+++ translate.lua
@@ -19,6 +19,10 @@ local options = {
     [[Output file.]]
   },
   {
+    '-save_attention', '',
+    [[Optional attention output file.]]
+  },
+  {
     '-batch_size', 30,
     [[Batch size.]],
     {
@@ -67,6 +71,7 @@ local function main()
   local goldBatch
 
   local withGoldScore = opt.tgt:len() > 0
+  local withAttention = opt.save_attention:len() > 0
 
   if withGoldScore then
     goldReader = onmt.utils.FileReader.new(opt.tgt, opt.idx_files)
@@ -74,6 +79,10 @@ local function main()
   end
 
   local outFile = io.open(opt.output, 'w')
+  local attFile
+  if withAttention then
+    attFile = io.open(opt.save_attention, 'w')
+  end
 
   local sentId = 1
   local batchId = 1
@@ -143,6 +152,28 @@ local function main()
             for n = 1, #results[b].preds do
               local sentence = translator:buildOutput(results[b].preds[n])
               outFile:write(sentence .. '\n')
+              if withAttention then
+                local attentions = results[b].preds[n].attention
+                for k,v in pairs(attentions) do
+                  if v ~= nil then
+                    local ttable = torch.totable(v)
+                    local attcount = #attentions
+                    local source = translator:buildOutput(srcBatch[b])
+                    local _,nt = sentence:gsub("%S+","")
+                    local _,ns = source:gsub("%S+","")
+                    if k == 1 then
+                      attFile:write(b..' ||| '..sentence..' ||| '..results[b].preds[n].score..' ||| '..source..' ||| '..ns..' '..nt..'\n')
+                    end
+                    for _,j in pairs(ttable) do
+                      attFile:write(j..' ')
+                    end
+                    attFile:write('\n')
+                    if k == attcount then
+                      attFile:write('\n')
+                    end
+                  end
+                end
+              end
               if n == 1 then
                 predScoreTotal = predScoreTotal + results[b].preds[n].score
                 predWordsTotal = predWordsTotal + #results[b].preds[n].words